Refactoring namespaces
