+++ src/controllers/GameOverManager.cs
@@ -1,6 +1,6 @@
 using Godot;
 
-namespace DiceRolling.Menus;
+namespace DiceRolling.Controllers;
 
 public partial class GameOverManager : Node2D {
     public override void _Ready() {
