+++ src/controllers/MainMenuManager.cs
@@ -1,6 +1,6 @@
 using Godot;
 
-namespace DiceRolling.Menus;
+namespace DiceRolling.MenuControllers;
 
 public partial class MainMenuManager : Node2D {
     public override void _Ready() {
