+++ src/models/stores/DiceEnergyStore.cs
@@ -1,5 +1,8 @@
 using Godot;
-namespace DiceRolling.Dice;
+
+using DiceRolling.Dice;
+
+namespace DiceRolling.Stores;
 
 [Tool]
 [GlobalClass]
