+++ src/renderers/scenes/gameplay/Dungeon.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=3 format=3 uid="uid://cbiewt2yq1klj"]
 
-[ext_resource type="Script" uid="uid://245qm3j5dd5t" path="res://models/data/Dungeon/DungeonManager.cs" id="1_5cacr"]
+[ext_resource type="Script" uid="uid://245qm3j5dd5t" path="res://controllers/DungeonManager.cs" id="1_5cacr"]
 [ext_resource type="Script" uid="uid://dpbxdw7s7l3mh" path="res://renderers/ui/SceneTransitionButton.cs" id="2_p0c6y"]
 
 [node name="Dungeon" type="Node2D"]
